Identity és SQLite szolgáltatások hozzáadása

+++ FoldrengesekSOE2026/ViewModels/Feladat4ViewModel.cs
@@ -5,6 +5,6 @@
         public string Nev { get; set; } = null!;
         public DateTime Datum { get; set; }
         public TimeSpan Ido { get; set; }
-        public decimal Magnitudo { get; set; }
+        public double Magnitudo { get; set; }
     }
 }
